requirements.txt
opencv-python>=4.5.5.64
numpy>=1.21.0
pyttsx3>=2.90
pyaudio>=0.2.11
webrtcvad>=2.0.10
openai-whisper>=20231117
torch>=2.0.0
transformers>=4.40.0
sentencepiece>=0.1.99